+++ src/pages/kyoungjinLee/Main/Main.js
@@ -1,6 +1,6 @@
-import { useState } from 'react';
+import { useEffect, useState } from 'react';
 import { Footer } from './footer';
-import CommentList from './Commentlist';
+import Feed from './feed';
 import './Main.scss';
 import '../../../styles/common.scss';
 
@@ -26,101 +26,6 @@ function Nav() {
 }
 
 //위스타 왼쪽
-
-//왼쪽 피드 프로필 ~
-function Feed({ commentList, userId }) {
-  return (
-    <>
-      {/*왼쪽 프로필*/}
-      <div class="meProfile">
-        <img
-          class="profileImage"
-          alt="내 프로필 이미지"
-          src="/images/kyoungjinLee/나.jpg"
-        />
-        <div class="meId">
-          <span class="meName">al_zar_takkarsen</span>
-          <span class="meDesc">🫠</span>
-        </div>
-        <div class="dots" />
-      </div>
-      {/*피드 콘텐츠*/}
-      <img alt="인서타 피드 이미지" src="/images/kyoungjinLee/싸콘.jpg" />
-      {/*피드 아래 아이콘들*/}
-      <div class="bottomIcon">
-        <button class="iconButton">
-          <img
-            alt="하트 아이콘"
-            class="icons like"
-            src="/images/kyoungjinLee/heart.png"
-          />
-        </button>
-        <img
-          alt="댓글 아이콘"
-          class="icons comment"
-          src="/images/kyoungjinLee/comments-2.png"
-        />
-        <img
-          class="icons upload"
-          alt="업로드 아이콘"
-          src="/images/kyoungjinLee/upload.png"
-        />
-        <img
-          alt="스크랩 아이콘"
-          class="icons scrap"
-          src="/images/kyoungjinLee/스크랩.png"
-        />
-      </div>
-      {/*~가 좋아합니다*/}
-      <div class="likey">
-        <img
-          class="likeyimage"
-          alt="내 프로필 이미지"
-          src="/images/kyoungjinLee/나.jpg"
-        />
-        <span class="likeComment">
-          <span class="bold">al_zar_takkarsen</span>님
-          <span class="bold">외 10명</span>이 좋아합니다
-        </span>
-      </div>
-      {/*댓글 들어갈 공간*/}
-      <div class="feedComment">
-        <span class="bold">al_zar_takkarsen</span> 우 나 핫걸 좋은 거 보고 갈래?
-        💦
-        <div class="comment1">
-          <span class="bold">_sentiers</span> 라인이 예술이네요
-          <button id="댓글좋아요" class="commentHeart">
-            <img
-              alt="좋아요 하트"
-              class="smallHeart"
-              src="/images/kyoungjinLee/heart.png"
-            />
-          </button>
-          <CommentList commentList={commentList} userId={userId} />
-        </div>
-        <div class="time">55분전</div>
-      </div>
-    </>
-  );
-}
-
-//댓글 쓰는 공간
-function CommentBox({ comment, getComment, addComment }) {
-  return (
-    <div class="getComment">
-      <input
-        id="댓글창"
-        type="text"
-        placeholder="댓글 달기..."
-        value={comment}
-        onChange={getComment}
-      />
-      <button id="게시" onClick={addComment}>
-        게시
-      </button>
-    </div>
-  );
-}
 
 //오른쪽 피드
 function MainRight({ footer }) {
@@ -222,18 +127,13 @@ function MainRight({ footer }) {
 }
 
 const MainKyoungjin = () => {
-  const [comment, setComment] = useState('');
-  const [commentList, setCommentList] = useState([]);
-  const userId = 'wak_good';
+  const [feedInfo, setFeedInfo] = useState([]);
 
-  const getComment = event => {
-    setComment(event.target.value);
-  };
-
-  const addComment = () => {
-    setCommentList([...commentList, comment]);
-    setComment('');
-  };
+  useEffect(() => {
+    fetch('/data/feed.json')
+      .then(res => res.json())
+      .then(data => setFeedInfo(data));
+  }, []);
 
   return (
     <>
@@ -242,14 +142,9 @@ const MainKyoungjin = () => {
         {/*왼쪽*/}
         <div class="feeds">
           <article>
-            <div class="meFid">
-              <Feed commentList={commentList} userId={userId} />
-              <CommentBox
-                comment={comment}
-                getComment={getComment}
-                addComment={addComment}
-              />
-            </div>
+            {feedInfo.map(content => (
+              <Feed content={content} key={content.id} />
+            ))}
           </article>
         </div>
         {/*오른쪽*/}
